Updated HTML and CSS
made QOL to the HTML and CSS. I made it so that the homepage has a different css than the other pages. I have also fixed the homepage text so that the animation happens as I wanted it to which was one text fades in at a time. Also made it so that you had to click the homepage image to play the snakegame as I kinda wanted it to be a secret.