Updated project vue, html and css
made it so that the vue list of projects contains the date and languages used for the project. header now has its own seperate css from the homepage fading in css. Lastly, made it so that project html shows the date. It gets the date from the list, uses the formatdate function in the vue code to make it more readable and puts it in the top right corner of the project box.